Executes git pull command and writes output to a file
I wrote my first piece of functionality, and installed RSpec so I could TDD it.

`git pull` and writes the output to a text file.
The file name has the current date in integer format appended to it,
so that the order in which the files were generated can be recreated.
I plan to chronologically parse the files and use the to create records in postgres.