Bug fix - lookup files based on their paths relative to the current directory
Ensures that the relative project path is used to determine a file's total line length and run flog on it. This fixes a bug I experienced wherein the all of the changed file records' total line lengths and flog scores were being set to nil. This happened because the file_path being used to calculate those values was not relative to the current directory, so the files could not be found. This doesn't happen in my tests because, in the specs, all file paths refer to files in the same project. In the real world, the path to the file from a directory outside a project will differ from the path to the file within that project (obviously).

To solve this, I load the environment variables in GitPullParser and prepend ENV["PATH_TO_PROJECT"] to the file_path within the project. The specs for this were ugly.