I made the relative file path specs less gross, by a lot!
This was ingenious of me the other day. I created a fixture ruby file against which to run flog, and then referenced that ruby file's path in the output_file_path fixture in the flog specs. This enabled me to test flog scores and get a consistent result on a file in the project that will not change, since it itself is a fixture. In other words, I created a way to call GitPullParser#parse_from_file only using fixtures, with no stubbing of git pull output, and no stubbing of flog score output.

Today, after writing a bunch of gross specs that relied on stubbing calls to `File#exist?` and other internals that I don't want to stub, I took a clever approach to getting rid of that stubbing, which was inspired by the above solution to my earlier problem. Even with the new approach, the one thing I still need to stub is the GitPullParser#path_to_project method, which returns the environment variable "PATH_TO_PROJECT", but that seems acceptable since the thing being stubbed is not part of the code, but rather the environment.

So what I did was: (1) set the stub of PATH_TO_PROJECT to "spec/fixtures", (2) create new git pull fixtures that remove "spec/fixtures" from the file paths of the changed files in the numstat output, and (3) check that "spec/fixtures" is prepended to the amended file paths in the spec expectations. The cleverness is that I am pretending that the first segment of a local file path is actually a relative project path, and thus ensuring the files exist, because they are real fixtures.